requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
berserk-downstream==0.11.4
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.12
discord==1.0.1
discord.py==1.6.0
idna==2.10
multidict==5.1.0
ndjson==0.3.1
requests==2.25.1
selenium==3.141.0
typing-extensions==3.7.4.3
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
